the typing problem is now fixed, but the reconstruction still blows up. This is a job for Friday

+++ ptypy/gpu/__init__.py
@@ -4,4 +4,4 @@ A module for gpu acceleration
 '''
 import numpy as np
 COMPLEX_TYPE= np.complex64
-FLOAT_TYPE = np.float64
+FLOAT_TYPE = np.float32
